more on types and casts
git-svn-id: https://svn.r-project.org/R/trunk@59166 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/RNG.c
@@ -163,7 +163,7 @@ double unif_rand(void)
 	II(0) = II(1); II(1) = II(2); II(2) = (int) p1;
 
 	p2 = a21 * (unsigned int)II(5) - a23n * (unsigned int)II(3);
-	k = p2 / m2;
+	k = (int) (p2 / m2);
 	p2 -= k * m2;
 	if (p2 < 0.0) p2 += m2;
 	II(3) = II(4); II(4) = II(5); II(5) = (int) p2;
